+++ src/components/index.ts
@@ -16,4 +16,4 @@ export * from "./slideOverFilters";
 export * from "./checkboxInput";
 export * from "./tableHeader";
 export * from "./tooltip";
-export * from "./AlertMessage";
+export * from "./alertMessage";
